get trajectory for selected cell set
