+++ python/src/worker/helpers/s3.py
@@ -43,6 +43,41 @@ def get_cell_sets(experiment_id):
 
         return cell_sets["cellSets"]
 
+def get_cell_sets_dict(cell_sets):
+    cell_sets_dict = {}
+
+    for cell_class in cell_sets:
+        if not cell_sets_dict.get(cell_class["key"]):
+          cell_sets_dict[cell_class["key"]] = {
+              "key": cell_class["key"],
+              "name": cell_class["name"],
+              "rootNode": True,
+              "childrenKeys": [],
+          }
+
+        for cell_set in cell_class["children"]:
+            cell_sets_dict[cell_class["key"]]["childrenKeys"].append(cell_set["key"])
+            cell_sets_dict[cell_set["key"]] = cell_set
+
+    return cell_sets_dict
+
+
+# Get all cell sets that match the subset_keys
+def get_cell_ids(subset_keys, cell_sets_dict):
+    cell_ids = set()
+
+    for subset_key in subset_keys:
+
+        # If cell class, get cell ids from child keys
+        if cell_sets_dict[subset_key]["rootNode"]:
+            cell_ids = cell_ids.union(
+                get_cell_ids(cell_sets_dict[subset_key]["childrenKeys"], cell_sets_dict)
+            )
+        else:
+          cell_ids = cell_ids.union(set(cell_sets_dict[subset_key]['cellIds']))
+
+    return cell_ids
+
 
 def get_embedding(etag, format_for_r):
     dir_path = os.path.join(config.LOCAL_DIR, f"{etag}")
