+++ r/R/trajectory_analysis.R
@@ -5,6 +5,8 @@
     umap = "UMAP"
   )
 
+  TRAJECTORY_ANALYSIS_CACHE <- list()
+
 #' Generate node coordinates for the initial trajectory analysis plot
 #'
 #' Returns a list containing node coordinates to the python worker
@@ -38,6 +40,7 @@ runTrajectoryAnalysisStartingNodesTask <- function(req, data) {
     req$body$embedding,
     req$body$embedding_settings,
     req$body$clustering_settings,
+    req$body$cell_ids,
     data
   )
 
@@ -119,28 +122,27 @@ runTrajectoryAnalysisPseudoTimeTask <- function(req, data) {
     req$body$embedding,
     req$body$embedding_settings,
     req$body$clustering_settings,
+    req$body$cell_ids,
     data
   )
 
   seurat_embedding_method <- req$body$embedding_settings$method
   monocle_embedding_method <- SEURAT_TO_MONOCLE_METHOD_MAP[[seurat_embedding_method]]
 
-
   node_ids <- colnames(cell_data@principal_graph_aux[[monocle_embedding_method]]$dp_mst)
 
   # Add 1 to indexes so they are 1-based
   root_indexes <- req$body$root_nodes + 1
-  
+
   # Translate the indexes to their ids
   root_ids <- node_ids[root_indexes]
-  
+
   cell_data <- monocle3::order_cells(cell_data, reduction_method = monocle_embedding_method, root_pr_nodes = root_ids)
 
   pseudotime <- as.data.frame(cell_data@principal_graph_aux@listData$UMAP$pseudotime)
 
   # fill in the NULL values for filtered cells
-  pseudotime <- fillNullForFilteredCells(pseudotime, data)
-
+  pseudotime <- fillNullForFilteredCells(pseudotime, subsetIds(data, req$body$cell_ids))
   result <- list(pseudotime = pseudotime[[1]])
   return(result)
 }
@@ -156,8 +158,28 @@ runTrajectoryAnalysisPseudoTimeTask <- function(req, data) {
 #'
 #' @return a cell_data_set object with cluster and graph information stored internally
 #' @export
-generateTrajectoryGraph <- function(embedding_data, embedding_settings, clustering_settings, data) {
+generateTrajectoryGraph <- function(
+  embedding_data,
+  embedding_settings,
+  clustering_settings,
+  cell_ids,
+  data
+) {
+
   set.seed(ULTIMATE_SEED)
+
+  cache_hash <- digest::digest(
+    list(
+      embedding_settings,
+      clustering_settings,
+      cell_ids
+    )
+  )
+
+  if(rlang::has_name(TRAJECTORY_ANALYSIS_CACHE, cache_hash)) {
+    message(paste0("Using hash with key ", cache_hash))
+    return(TRAJECTORY_ANALYSIS_CACHE[[cache_hash]])
+  }
 
   Seurat::DefaultAssay(data) <- "RNA"
 
@@ -178,6 +200,7 @@ generateTrajectoryGraph <- function(embedding_data, embedding_settings, clusteri
   }
 
   data <- assignEmbedding(embedding_data, data)
+  data <- subsetIds(data, cell_ids)
 
   cell_data <- SeuratWrappers::as.cell_data_set(data)
 
@@ -192,6 +215,10 @@ generateTrajectoryGraph <- function(embedding_data, embedding_settings, clusteri
   )
 
   cell_data <- monocle3::learn_graph(cell_data)
+
+  TRAJECTORY_ANALYSIS_CACHE[[cache_hash]] <- cell_data
+
+  message(paste0("Trajectory graph cell data cached with key ", cache_hash))
 
   return(cell_data)
 }
